counting towers

+++ DP/1_dice.cpp
@@ -7,7 +7,7 @@ ll helper(ll n){
     return helper(n-1) + helper(n-2) + helper(n-3) + helper(n-4) + helper(n-5) + helper(n-6);
 }
 ll func(ll n){
-    int MOD = 1e9+7;
+    ll MOD = 1e9+7;
     vector<ll> dp(n+1,0);
     dp[0] = 1;
     if(n == 0) return dp[n];
@@ -21,7 +21,7 @@ ll func(ll n){
     if(n == 4) return dp[n];
     dp[5] = dp[0] + dp[1] + dp[2] + dp[3] + dp[4];
     if(n == 5) return dp[n];
-    for(int i = 6;i<=n;i++){
+    for(ll i = 6;i<=n;i++){
         dp[i] = (dp[i] + dp[i-1])%MOD;
         dp[i] = (dp[i] + dp[i-2])%MOD;
         dp[i] = (dp[i] + dp[i-3])%MOD;
