+++ DP/2_minimize_coins.cpp
@@ -1,7 +1,7 @@
 #include<bits/stdc++.h>
 using namespace std;
 #define ll long long
-ll memo(int n,int sum,vector<ll>&arr){
+ll memo(ll n,ll sum,vector<ll>&arr){
     if(sum == 0) return 0;
     if(n == 0) return 1e9;
     ll pick = 1e9;
@@ -12,8 +12,8 @@ ll memo(int n,int sum,vector<ll>&arr){
 ll helper(vector<ll>&arr,ll n,ll sum){
     vector<ll> dp(sum+1,1e9);
     dp[0] = 0;
-    for(int i = 1;i<=sum;i++){
-        for(int j = n-1;j>=0;j--){
+    for(ll i = 1;i<=sum;i++){
+        for(ll j = n-1;j>=0;j--){
             if(i >= arr[j]) dp[i] = min(dp[i],1+dp[i-arr[j]]);
         }
     }
@@ -24,7 +24,7 @@ int main(){
     ll n,k;
     cin>>n>>k;
     vector<ll> v(n);
-    for(int i = 0;i<n;i++){
+    for(ll i = 0;i<n;i++){
         cin>>v[i];
     }
     cout<<helper(v,n,k)<<endl;
