+++ Introductory Problems/5_permutations.cpp
@@ -2,8 +2,8 @@
 using namespace std;
 int main(){
 
-    // Logic is that first print all odd one by one and then even. there difference is always more than one.
-    int n;
+    // Logic is that first prll all odd one by one and then even. there difference is always more than one.
+    ll n;
     cin>>n;
     if(n == 1){
         cout<<"1";
@@ -13,10 +13,10 @@ int main(){
         cout<<"NO SOLUTION"<<endl;
         return 0;
     }
-    for(int i = 2;i<=n;i = i+2){
+    for(ll i = 2;i<=n;i = i+2){
         cout<<i<<" ";
     }
-    for(int i = 1;i<=n;i = i+2){
+    for(ll i = 1;i<=n;i = i+2){
         if(i == n || i == n-1){
             cout<<i;
         }else{
